+++ shared/opts-parser.js
@@ -64,6 +64,9 @@ var OptsParser;
         server: {
             param: 'string', desc: "name of the server. default is " + rel_consts_js_1.RelConsts.DEFAULT_SERVER,
         },
+        serverExec: {
+            param: 'string', desc: "name of the server executable. default puppeteer=node. slimerjs=slimerjs",
+        },
         fps: {
             param: 'int', desc: "overrides frames per second defined on project",
         },